Extract config, scheduler and gc parts from waymark crate
