+++ Cargo.toml
@@ -15,6 +15,8 @@ waymark-backends-core = { path = "crates/lib/backends-core" }
 waymark-core-backend = { path = "crates/lib/core-backend" }
 waymark-dag = { path = "crates/lib/dag" }
 waymark-garbage-collector-backend = { path = "crates/lib/garbage-collector-backend" }
+waymark-garbage-collector-config = { path = "crates/lib/garbage-collector-config" }
+waymark-garbage-collector-loop = { path = "crates/lib/garbage-collector-loop" }
 waymark-support-integration = { path = "crates/lib/support-integration" }
 waymark-ir-conversions = { path = "crates/lib/ir-conversions" }
 waymark-ir-parser = { path = "crates/lib/ir-parser" }
@@ -26,6 +28,8 @@ waymark-runner = { path = "crates/lib/runner" }
 waymark-runner-state = { path = "crates/lib/runner-state" }
 waymark-scheduler-backend = { path = "crates/lib/scheduler-backend" }
 waymark-scheduler-core = { path = "crates/lib/scheduler-core" }
+waymark-scheduler-config = { path = "crates/lib/scheduler-config" }
+waymark-scheduler-loop = { path = "crates/lib/scheduler-loop" }
 waymark-support-test = { path = "crates/lib/support-test" }
 waymark-webapp-backend = { path = "crates/lib/webapp-backend" }
 waymark-webapp-core = { path = "crates/lib/webapp-core" }
@@ -44,6 +48,7 @@ waymark-worker-status-core = { path = "crates/lib/worker-status-core" }
 waymark-worker-status-reporter = { path = "crates/lib/worker-status-reporter" }
 waymark-worker-metrics = { path = "crates/lib/worker-metrics" }
 waymark-runloop = { path = "crates/lib/runloop" }
+waymark-config = { path = "crates/lib/config" }
 
 anyhow = "1"
 async-trait = "0.1"
