+++ crates/lib/garbage-collector-config/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "waymark-garbage-collector-config"
+edition = "2024"
+version.workspace = true
+
+[dependencies]
