+++ crates/waymark/Cargo.toml
@@ -5,31 +5,10 @@ version.workspace = true
 edition = "2024"
 
 [dependencies]
-waymark-core-backend = { workspace = true }
 waymark-dag = { workspace = true }
-waymark-ir-conversions = { workspace = true }
 waymark-ir-parser = { workspace = true }
 waymark-proto = { workspace = true, features = ["serde", "client", "server"] }
-waymark-runner = { workspace = true }
-waymark-runner-state = { workspace = true }
-waymark-garbage-collector-backend = { workspace = true }
-waymark-scheduler-backend = { workspace = true }
-waymark-scheduler-core = { workspace = true }
-waymark-backends-core = { workspace = true }
-waymark-backend-memory = { workspace = true }
-waymark-message-conversions = { workspace = true }
-waymark-webapp-config = { workspace = true }
-
-anyhow = "1"
-prost = "0.12"
 serde_json = "1"
-chrono = { version = "0.4", default-features = false, features = ["clock", "serde"] }
-uuid = { version = "1", features = ["serde", "v4"] }
 thiserror = { workspace = true }
-tokio = { version = "1", features = ["full"] }
-tokio-util = "0.7"
-tonic = { version = "0.11", features = ["transport"] }
-tracing = "0.1"
 
 [dev-dependencies]
-waymark-backend-memory = { workspace = true }
