--- crates/waymark/src/garbage_collector/mod.rs
@@ -1,5 +1,0 @@
-//! Background garbage collection for old finished workflow instances.
-
-mod task;
-
-pub use task::{GarbageCollectorConfig, GarbageCollectorTask};
