--- crates/waymark/src/scheduler/mod.rs
@@ -1,10 +1,0 @@
-//! Workflow scheduling system.
-//!
-//! This module provides:
-//! - Schedule types and database operations
-//! - Background task for firing due schedules
-//! - Cron and interval utilities
-
-mod task;
-
-pub use task::{DagResolver, SchedulerConfig, SchedulerTask, WorkflowDag};
